requirements.txt
requests==2.31.0
urllib3==1.26.15
python-dotenv==1.0.0
fake-useragent==1.2.1
goose3==3.1.13
beautifulsoup4==4.12.2
selenium==4.10.0
splinter==0.18.1
webdriver-manager==4.0.0